refactor: generic callback manager
